Exception handling added

+++ src/main/java/com/ing/testapp/store/StoreController.java
@@ -1,5 +1,7 @@
 package com.ing.testapp.store;
 
+import com.ing.testapp.exceptions.InvalidProductIdException;
+import com.ing.testapp.exceptions.ProductIdNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -16,16 +18,31 @@ public class StoreController {
     public ResponseEntity<String> addProduct(
             @RequestBody Product newProduct
     ) {
-        productRepository.save(newProduct);
-        return new ResponseEntity<>("New product added successfully.", HttpStatus.CREATED);
+        try {
+            productRepository.save(newProduct);
+            return new ResponseEntity<>("New product added successfully.", HttpStatus.CREATED);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Error encountered on add operation.", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @DeleteMapping(path = "/delete_product")
     public ResponseEntity<String> deleteProduct(
             @RequestParam Integer id
     ) {
-        productRepository.deleteById(id);
-        return new ResponseEntity<>(String.format("Product with id: %d deleted successfully.", id), HttpStatus.OK);
+        if (id <= 0) {
+            throw new InvalidProductIdException();
+        }
+        if (!productRepository.existsById(id)) {
+            throw new ProductIdNotFoundException();
+        }
+
+        try {
+            productRepository.deleteById(id);
+            return new ResponseEntity<>(String.format("Product with id: %d deleted successfully.", id), HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Error encountered on delete operation.", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @PutMapping(path = "/update_product", consumes = MediaType.APPLICATION_JSON_VALUE)
@@ -33,6 +50,11 @@ public class StoreController {
             @RequestParam Integer id,
             @RequestBody Product newProduct
     ) {
+        if (id <= 0) {
+            throw new InvalidProductIdException();
+        }
+
+        try {
         Product updatedProduct = productRepository.findById(id)
                 .map(product -> {
                     product.setName(newProduct.getName());
@@ -43,26 +65,48 @@ public class StoreController {
                 })
                 .orElseGet(() -> newProduct);
 
-        productRepository.save(updatedProduct);
-        return new ResponseEntity<>(String.format("Product with id: %d updated successfully.", id), HttpStatus.OK);
+            productRepository.save(updatedProduct);
+            return new ResponseEntity<>(String.format("Product with id: %d updated successfully.", id), HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Error encountered on update operation.", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @GetMapping(path = "/find_colour")
-    public @ResponseBody Iterable<Product> getProductsByColour(
+    public ResponseEntity<Object> getProductsByColour(
             @RequestParam String colour
     ) {
-        return productRepository.findByColour(colour);
+        try {
+            return new ResponseEntity(productRepository.findByColour(colour), HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Error encountered on find operation.", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @GetMapping(path = "/find_product")
-    public @ResponseBody Product getProductById(
+    public ResponseEntity<Object> getProductById(
             @RequestParam Integer id
     ) {
-        return productRepository.findById(id).get();
+        if (id <= 0) {
+            throw new InvalidProductIdException();
+        }
+        if (!productRepository.existsById(id)) {
+            throw new ProductIdNotFoundException();
+        }
+
+        try {
+            return new ResponseEntity(productRepository.findById(id).get(), HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Error encountered on find operation.", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @GetMapping(path = "/all_products")
-    public @ResponseBody Iterable<Product> getAllProducts() {
-        return productRepository.findAll();
+    public ResponseEntity<Object> getAllProducts() {
+        try {
+            return new ResponseEntity(productRepository.findAll(), HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>("Error encountered on find operation.", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
